Fix sporadic test failure
One of the tests in test_pstreams.cc assumes that two adjacent 'date'
commands will produce the same output, but that's not true if the clock
happens to tick to a new second between the commands. This is very
unlikely to happen when run interactively, but is more likely on CI
systems that are slow, heavily loaded, shared by many processes, or all
of the above. Change the test to allow one second differences, but also
print out the times so that the logs will show why it failed if for some
reason the process stalls for more than a second.